requirements.txt
fastapi
uvicorn[standard]
openai>=1.0.0
pydantic
python-dotenv
openai-agents